Fix: Added ASCII85 library loading check in convertText function
In this commit, I have added a check for the successful loading of the ASCII85 library before attempting to use it in the convertText function. If the library is not loaded, it will display an error message in the ASCII85 output textarea. This ensures that the rest of the script continues to work even if there is an issue with loading the library.